refactor(architecture): move changelog into release domain
