+++ src/core/mod.rs
@@ -4,7 +4,6 @@ pub mod config;
 pub mod api;
 pub mod auth;
 pub mod build;
-pub mod changelog;
 pub mod cleanup;
 pub mod cli_tool;
 pub mod code_audit;
